Emit statistics from puzzle solver.
The Puzzle type can optionally be configured to use a profiler by passing the `Profile` or `NoProfile` type parameter. If `NoProfile` is chosen, no runtime cost is incurred.

Currently, the following statistics are tracked:
* Maximum depth of tree.
* Number of nodes in the tree.

The unit tests for the puzzles also assert on these statistics to give an additional way of tracking changes in performance.

Closes #14.